Extend set definition data with genes on the same row separated by commas, semicolons, slashes, etc

Extend set definition data with genes on the same row separated by commas, semicolons, slashes, etc

Usage

extend_set_data(set.data, sep)

Arguments

set.data

A data frame with set definition data.

sep

Punctuation to split on.

Value

A data frame.

Examples

data(setdat)
extend_set_data(setdat, sep=",")
#>             set.id mtx.id                   row.id
#> 1  ENSG00000121410    RNA          ENSG00000121410
#> 2  ENSG00000148584    RNA          ENSG00000148584
#> 3  ENSG00000175899    RNA          ENSG00000175899
#> 4  ENSG00000166535    RNA          ENSG00000166535
#> 5  ENSG00000184389    RNA          ENSG00000184389
#> 6  ENSG00000128274    RNA          ENSG00000128274
#> 7  ENSG00000118017    RNA          ENSG00000118017
#> 8  ENSG00000094914    RNA          ENSG00000094914
#> 9  ENSG00000081760    RNA          ENSG00000081760
#> 10 ENSG00000114771    RNA          ENSG00000114771
#> 11 ENSG00000197953    RNA          ENSG00000197953
#> 12 ENSG00000188984    RNA          ENSG00000188984
#> 13 ENSG00000204518    RNA          ENSG00000204518
#> 14 ENSG00000109576    RNA          ENSG00000109576
#> 15 ENSG00000224854    RNA          ENSG00000224854
#> 16 ENSG00000147889    RNA          ENSG00000147889
#> 17 ENSG00000147883    RNA          ENSG00000147883
#> 18 ENSG00000176399    RNA          ENSG00000176399
#> 19 ENSG00000099810    RNA          ENSG00000099810
#> 20 ENSG00000148400    RNA          ENSG00000148400
#> 21 ENSG00000264545 Lesion     ENSG00000264545_loss
#> 22 ENSG00000147889 Lesion     ENSG00000147889_loss
#> 23 ENSG00000224854 Lesion     ENSG00000224854_loss
#> 24 ENSG00000148400 Lesion ENSG00000148400_mutation
#> 25 ENSG00000099810 Lesion     ENSG00000099810_loss
#> 26 ENSG00000264801 Lesion     ENSG00000264801_loss
#> 27 ENSG00000240498 Lesion     ENSG00000240498_loss
#> 28 ENSG00000265194 Lesion     ENSG00000265194_loss
#> 29 ENSG00000266446 Lesion     ENSG00000266446_loss
#> 30 ENSG00000147883 Lesion     ENSG00000147883_loss
#> 31 ENSG00000215221 Lesion     ENSG00000215221_loss
#> 32 ENSG00000266732 Lesion     ENSG00000266732_loss
#> 33 ENSG00000236921 Lesion     ENSG00000236921_loss
#> 34 ENSG00000229298 Lesion     ENSG00000229298_loss
#> 35 ENSG00000244230 Lesion     ENSG00000244230_loss
#> 36 ENSG00000229835 Lesion     ENSG00000229835_loss
#> 37 ENSG00000176399 Lesion     ENSG00000176399_loss
#> 38 ENSG00000227443 Lesion     ENSG00000227443_loss
#> 39 ENSG00000171889 Lesion     ENSG00000171889_loss
#> 40 ENSG00000199177 Lesion     ENSG00000199177_loss